Fixed minor bugs

+++ app/src/main/java/android/example/com/currencycalculator/api/CurrencyCalculatorApi.kt
@@ -8,5 +8,5 @@ import retrofit2.http.Query
 interface CurrencyCalculatorApi {
 
     @GET("fixer/convert")
-    suspend fun getFixerConvert(@Query("to") to: String, @Query("from") from: String, @Query("amount") amount: Int): Response<Fixer>
+    suspend fun getFixerConvert(@Query("to") to: String, @Query("from") from: String, @Query("amount") amount: Float): Response<Fixer>
 }
