+++ app/src/main/java/android/example/com/currencycalculator/repository/CurrencyCalculatorRepository.kt
@@ -4,6 +4,6 @@ import android.example.com.currencycalculator.api.authentication.RetrofitInstanc
 
 class CurrencyCalculatorRepository() {
 
-    suspend fun getFixerConvert(to: String, from: String, amount: Int) = RetrofitInstance.api.getFixerConvert(to, from, amount)
+    suspend fun getFixerConvert(to: String, from: String, amount: Float) = RetrofitInstance.api.getFixerConvert(to, from, amount)
 
 }
